requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
httpx==0.28.1
pandas==2.2.3
numpy==2.2.5
lifelines==0.30.0
openai==1.76.2
python-dotenv==1.1.0
